Copywriter fix

+++ repository/CNV_RunCopywriter.R
@@ -46,7 +46,7 @@ resolution=as.numeric(as.character(resolution))/1000
 destination.folder = file.path(paste(name,"/results/Copywriter/",sep=""))
 
 # Copywriter will break if the destination folder already has the folders from a previous run (this is only an issue for rerunning Copywriter)
-dir.exists(paste0(destination.folder, "/CNAprofiles")){
+if(dir.exists(paste0(destination.folder, "/CNAprofiles"))){
   unlink(paste0(destination.folder, "/CNAprofiles"))
 }
